requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
channels[daphne]==4.0.0
channels-redis==4.0.0